Java demo project
