+++ examples/java/hello-secret/.gitattributes
