+++ examples/java/hello-secret/.gitignore
@@ -3,3 +3,4 @@
 
 # Ignore Gradle build output directory
 build
+/config*
