+++ examples/java/hello-secret/gradle/wrapper/gradle-wrapper.properties
