+++ examples/java/hello-secret/settings.gradle
