+++ examples/java/hello-secret/src/test/java/internal/AppTest.java
