--- sdk/java/internal/src/main/java/internal/App.java
@@ -1,15 +1,0 @@
-package internal;
-
-import java.io.File;
-
-import static com.keepersecurity.secretsManager.core.SecretsManager.*;
-
-public class App {
-
-    public File configFile = new File("config.txt");
-
-    public static void main(String[] args) {
-
-//        getSecrets();
-    }
-}
